@@ -120,40 +118,37 @@
     "@babel/plugin-proposal-pipeline-operator": "7.18.9",
     "@babel/plugin-proposal-throw-expressions": "7.18.6",
     "@babel/plugin-syntax-import-meta": "7.10.4",
-    "@babel/preset-env": "7.19.3",
+    "@babel/preset-env": "7.20.2",
     "@babel/preset-react": "7.18.6",
     "@babel/preset-typescript": "7.18.6",
-    "@lavamoat/allow-scripts": "2.0.3",
+    "@lavamoat/allow-scripts": "2.1.0",
     "@lavamoat/preinstall-always-fail": "1.0.0",
-    "@ledgerhq/hw-transport-mocker": "6.27.5",
-    "@nomiclabs/hardhat-ethers": "^2.1.1",
-    "@nomiclabs/hardhat-waffle": "2.0.3",
-    "@sentry/types": "7.14.0",
-    "@testing-library/dom": "8.18.1",
+    "@ledgerhq/hw-transport-mocker": "6.27.7",
+    "@sentry/types": "7.20.1",
+    "@testing-library/dom": "8.19.0",
     "@testing-library/react": "13.4.0",
     "@testing-library/user-event": "14.4.3",
     "@types/cookie": "0.5.1",
     "@types/get-pixels": "3.3.2",
     "@types/hdkey": "2.0.1",
-    "@types/jest": "29.1.0",
-    "@types/node": "18.7.23",
+    "@types/jest": "29.2.3",
+    "@types/node": "18.11.9",
     "@types/node-fetch": "2.6.2",
     "@types/node-hid": "1.3.1",
     "@types/uuid": "8.3.4",
     "@types/ws": "8.5.3",
     "babel-eslint": "10.1.0",
-    "electron": "20.3.0",
-    "electron-builder": "23.3.3",
+    "electron": "21.3.0",
+    "electron-builder": "23.6.0",
     "electron-notarize": "1.2.1",
-    "eslint": "8.24.0",
-    "ethereum-waffle": "3.4.4",
-    "ethers": "5.7.1",
-    "hardhat": "2.11.2",
-    "jest": "29.1.1",
-    "jest-environment-jsdom": "29.1.1",
+    "eslint": "8.28.0",
+    "ethers": "5.7.2",
+    "hardhat": "2.12.2",
+    "jest": "29.3.1",
+    "jest-environment-jsdom": "29.3.1",
     "nock": "13.2.9",
-    "node-abi": "3.25.0",
-    "node-gyp": "9.1.0",
+    "node-abi": "3.28.0",
+    "node-gyp": "9.3.0",
     "node-watch": "0.7.3",
     "parcel-bundler": "1.12.5",
     "prebuild-install": "7.1.1",
@@ -162,7 +157,7 @@
     "standardx": "7.0.0",
     "stylus": "0.59.0",
     "ts-jest": "29.0.3",
-    "typescript": "4.8.4"
+    "typescript": "4.9.3"
   },
   "standardx": {
     "parser": "babel-eslint",
