@@ -1,13 +1,10 @@
-import React, { Children } from 'react'
+import React from 'react'
 import Restore from 'react-restore'
 import svg from '../../../../../../../resources/svg'
 import link from '../../../../../../../resources/link'
-import utils from 'web3-utils'
 import BigNumber from 'bignumber.js'
-import Transfer from './erc20/transfer'
 import Recipient from './recipient'
 import Destination from './destination'
-import Register from './ens/register'
 import { ClusterBox, Cluster, ClusterRow, ClusterValue } from '../../../../../../../resources/Components/Cluster'
 
 import { formatDisplayInteger, isUnlimited } from '../../../../../../../resources/utils/numbers'
@@ -24,9 +21,7 @@ class TxSending extends React.Component {
     this.setState({ copied: true })
     setTimeout(_ => this.setState({ copied: false }), 1000)
   }
-  hexToDisplayValue (hex) {
-    return (Math.round(parseFloat(utils.fromWei(hex, 'ether')) * 1000000) / 1000000).toFixed(6)
-  }
+  
   render () {
     const req = this.props.req
     const contract = req.data.to.toLowerCase()
