@@ -1,5 +1,5 @@
 // NPM modules
-const codec = require('abi-codec')
+const {utils: {Interface}} = require('ethers')
 const namehash = require('eth-ens-namehash')
 const contentHash = require('content-hash')
 
@@ -11,6 +11,9 @@ const store = require('../store').default
 const interfaces = require('./artifacts/interfaces')
 const registryAddresses = require('./artifacts/addresses')
 
+const registryContractInterface = new Interface(interfaces.registry)
+const resolverContractInterface = new Interface(interfaces.resolver)
+
 /* PUBLIC */
 exports.resolveName = async (name) => {
   // Get resolver address
@@ -21,7 +24,7 @@ exports.resolveName = async (name) => {
 
   // Encode function input
   const node = namehash.hash(name)
-  const input = codec.encodeInput(interfaces.resolver, 'addr', [node])
+  const input = resolverContractInterface.encodeFunctionData('addr', [node])
 
   // Make JSON RPC call
   const params = { to: resolverAddress, data: input }
@@ -31,7 +34,8 @@ exports.resolveName = async (name) => {
   if (output === '0x') return null
 
   // Decode output and return value
-  const decodedOutput = codec.decodeOutput(interfaces.resolver, 'addr', output)
+  const decodedOutput = resolverContractInterface.decodeFunctionResult('addr', output)
+  
   return decodedOutput[0]
 }
 
@@ -47,7 +51,7 @@ exports.resolveAddress = async (address) => {
 
   // Encode function input
   const node = namehash.hash(name)
-  const input = codec.encodeInput(interfaces.resolver, 'name', [node])
+  const input = resolverContractInterface.encodeFunctionData('name', [node])
 
   // Make JSON RPC call
   const params = { to: resolverAddress, data: input }
@@ -57,7 +61,7 @@ exports.resolveAddress = async (address) => {
   if (output === '0x') return null
 
   // Decode output and return value
-  const decodedOutput = codec.decodeOutput(interfaces.resolver, 'name', output)
+  const decodedOutput = resolverContractInterface.decodeFunctionResult('name', output)
   return decodedOutput[0]
 }
 
@@ -70,7 +74,7 @@ exports.resolveContent = async (name) => {
 
   // Encode function input
   const node = namehash.hash(name)
-  const input = codec.encodeInput(interfaces.resolver, 'contenthash', [node])
+  const input = resolverContractInterface.encodeFunctionData('contenthash', [node])
 
   // Make JSON RPC call
   const params = { to: resolverAddress, data: input }
@@ -80,7 +84,7 @@ exports.resolveContent = async (name) => {
   if (output === '0x') return null
 
   // Decode output and return the content hash in text format
-  const decodedOutput = codec.decodeOutput(interfaces.resolver, 'contenthash', output)
+  const decodedOutput = resolverContractInterface.decodeFunctionResult('contenthash', [node])
 
   if (decodedOutput[0] === null) return null
 
@@ -101,7 +105,7 @@ const getResolverAddress = async (name) => {
   const registryAddress = registryAddresses[networkId]
 
   // Encode function input
-  const input = codec.encodeInput(interfaces.registry, 'resolver', [hash])
+  const input = registryContractInterface.encodeFunctionData('resolver', [hash])
 
   // Make JSON RPC call
   const params = { to: registryAddress, data: input }
@@ -111,7 +115,7 @@ const getResolverAddress = async (name) => {
   if (output === '0x') return null
 
   // Decode output and return value
-  const decodedOutput = codec.decodeOutput(interfaces.registry, 'resolver', output)
+  const decodedOutput = registryContractInterface.decodeFunctionResult('resolver', output)
   return decodedOutput[0]
 }
 
