@@ -1,5 +1,5 @@
 import log from 'electron-log'
-import utils from 'web3-utils'
+import { hexToInt } from '../../../../resources/utils'
 import { padToEven, stripHexPrefix, addHexPrefix } from 'ethereumjs-util'
 import { TypedData, TypedDataUtils } from 'eth-sig-util'
 import type { Device as TrezorDevice } from 'trezor-connect'
@@ -308,7 +308,7 @@ export default class Trezor extends Signer {
       to: this.normalize(to || ''),
       value: this.normalize(value || ''),
       data: this.normalize(data || ''),
-      chainId: utils.hexToNumber(chainId)
+      chainId: hexToInt(chainId)
     }
 
     const optionalFields = ['gasPrice', 'maxFeePerGas', 'maxPriorityFeePerGas']
